requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
simpleeval>=0.9.13
